requirements.txt
streamlit
torch
torchvision
timm
facenet-pytorch
pillow
numpy
pandas
